"check_style_score -> 4 of 5 passed. style score: 0/1 "

+++ cspp2 assignment/m2/Assignment-3/Solution.java
@@ -19,7 +19,7 @@ public class Solution {
 		Scanner s = new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        int result = power(base,exponent);
+        long result = power(base,exponent);
         System.out.println(result);
 	}
 	/**
